ai-agent/requirements.txt
paho-mqtt==1.6.1
scikit-learn==1.2.2
pandas==2.0.0
python-telegram-bot==20.3
spacy==3.5.2
numpy==1.24.2
python-dotenv==1.0.0
